// src/enums/SectionEnum.ts
export enum SectionEnum {
  User = 'User',
  Dashboard = 'Dashboard',
  ManualControl = 'ManualControl',
  ProtocolExplorer = 'ProtocolExplorer',
  ReportExplorer = 'ReportExplorer',
  RegularRun = 'RegularRun',
  ProtocolSetting = 'ProtocolSetting',
  Stampede = 'Stampede',
  NotImplemented = 'NotImplemented',
  Network ='Network'
}